fix: move startable and events interfaces
The index file is a dumping ground for interfaces, move them into their
own files to keep things managable.

Also fixes pubsub test to actually shut the nodes down.
