+++ packages/libp2p-interface-compliance-tests/src/transport/listen-test.ts
@@ -9,7 +9,7 @@ import { mockUpgrader } from '../mocks/upgrader.js'
 import defer from 'p-defer'
 import { mockRegistrar } from '../mocks/registrar.js'
 import drain from 'it-drain'
-import { CustomEvent } from '@libp2p/interfaces'
+import { CustomEvent } from '@libp2p/interfaces/events'
 import type { TestSetup } from '../index.js'
 import type { Transport, Upgrader } from '@libp2p/interfaces/transport'
 import type { TransportTestFixtures } from './index.js'
