+++ packages/libp2p-interfaces/src/components.ts
@@ -1,7 +1,8 @@
 import errCode from 'err-code'
 import type { ConnectionGater, ConnectionProtector } from './connection/index.js'
 import type { ContentRouting } from './content-routing/index.js'
-import { AddressManager, isStartable, Startable } from './index.js'
+import type { AddressManager } from './address-manager/index.js'
+import { isStartable, Startable } from './startable.js'
 import type { Metrics } from './metrics/index.js'
 import type { PeerId } from './peer-id/index.js'
 import type { PeerRouting } from './peer-routing/index.js'
