+++ packages/libp2p-interfaces/src/peer-discovery/index.ts
@@ -1,5 +1,5 @@
 import type { PeerInfo } from '../peer-info/index.js'
-import type { EventEmitter } from '../index.js'
+import type { EventEmitter } from '../events.js'
 
 export interface PeerDiscoveryEvents {
   'peer': CustomEvent<PeerInfo>
