+++ packages/libp2p-interfaces/src/pubsub/index.ts
@@ -1,6 +1,6 @@
 import type { PeerId } from '../peer-id/index.js'
 import type { Pushable } from 'it-pushable'
-import type { EventEmitter } from '../index.js'
+import type { EventEmitter } from '../events.js'
 import type { Stream } from '../connection/index.js'
 
 /**
