+++ packages/libp2p-peer-store/src/address-book.ts
@@ -9,7 +9,7 @@ import filter from 'it-filter'
 import map from 'it-map'
 import each from 'it-foreach'
 import { peerIdFromPeerId } from '@libp2p/peer-id'
-import { CustomEvent } from '@libp2p/interfaces'
+import { CustomEvent } from '@libp2p/interfaces/events'
 import type { AddressFilter, Peer, PeerMultiaddrsChangeData, PeerStore } from '@libp2p/interfaces/peer-store'
 import type { Store } from './store.js'
 import type { Envelope } from '@libp2p/interfaces/record'
