+++ packages/libp2p-peer-store/src/key-book.ts
@@ -3,7 +3,7 @@ import errcode from 'err-code'
 import { codes } from './errors.js'
 import { peerIdFromPeerId } from '@libp2p/peer-id'
 import { equals as uint8arrayEquals } from 'uint8arrays/equals'
-import { CustomEvent } from '@libp2p/interfaces'
+import { CustomEvent } from '@libp2p/interfaces/events'
 import type { Store } from './store.js'
 import type { PeerStore, KeyBook, PeerPublicKeyChangeData, Peer } from '@libp2p/interfaces/peer-store'
 import type { PeerId } from '@libp2p/interfaces/peer-id'
