+++ packages/libp2p-peer-store/src/proto-book.ts
@@ -2,7 +2,7 @@ import { logger } from '@libp2p/logger'
 import errcode from 'err-code'
 import { codes } from './errors.js'
 import { peerIdFromPeerId } from '@libp2p/peer-id'
-import { CustomEvent } from '@libp2p/interfaces'
+import { CustomEvent } from '@libp2p/interfaces/events'
 import type { Store } from './store.js'
 import type { Peer, PeerProtocolsChangeData, PeerStore, ProtoBook } from '@libp2p/interfaces/peer-store'
 import type { PeerId } from '@libp2p/interfaces/peer-id'
